added exceptions to userservice

+++ Project_1/Gallery.API/3_Service/UserService.cs
@@ -1,3 +1,4 @@
+using Gallery.API.CustomExceptions;
 using Gallery.API.Model;
 using Gallery.API.Repository;
 
@@ -14,18 +15,49 @@ public class UserService : IUserService{
         return _userRepository.GetAllUsers();
     }
 
-    public User CreateNewUser(User user){        
+    public User CreateNewUser(User user){    
+
+       List<string> rolesList = ["admin", "artist", "guest"];
+       string roleInput = user.Role.ToLower();
+
+       if (!rolesList.Contains(roleInput))
+       {
+        throw new Exception("user_Role must be: Admin, Artist, or Guest");
+       }       
+
        return _userRepository.CreateNewUser(user);
     }
 
     public User? GetUserById(int id)
     {
-        
-        return _userRepository.GetUserById(id);
+        if (id<1)
+        {
+            throw new Exception("ID must be a positive integer!");
+        }
+
+        var u = _userRepository.GetUserById(id);
+
+        if (u==null)
+        {
+            throw new UserNotFoundException("This user doesn't exist!");
+        } 
+        return u;
     }
 
     public User DeleteUserById(int id)
     {
+        if (id<1)
+        {
+            throw new Exception("ID must be a positive integer!");
+        }
+
+        var u = _userRepository.GetUserById(id);
+
+        if (u==null)
+        {
+            throw new UserNotFoundException("This user doesn't exist!");
+        } 
+
         return _userRepository.DeleteUserById(id);
     }
 }
